fix: nasty! sometimes targets have duplicate labels
Because I forgot to update the code to use correct array size.
I reuse the array for storing one-char (initial) labels, and
because Lua is the language that it is (1. arrays are downsized, pointless
to clear the array and 2. no easy way; note: 1 might not apply to luajit
but who knows), I just left the junk from prev iterations at the end.
And the size of the array with that junk was
used instead of the valid size when computing the labels.